lava-qemu: Simplifications but remain working.
Simplified and made more like the parameters from qemu-run.sh
This critical item was -m 2048.  -m 3072 also worked so I went with that.
Using -m 4096 or even 3073 dies after the bl31 header message.

Also note that TRS will not run twice this way.
Multiple reboots in one run are fine.

However the tpm instance is different each time so when TRS encypts the
rootfs it stores the key in tpm but the next run has a different tpm so
it just reboots.

Signed-off-by: Bill Mills <bill.mills@linaro.org>