qemu-run.sh: reset TRS image each time
As the docker setup is right now, the TPM starts fresh each time.

TRS encrypts the rootfs to be unique to the TPM state. This works find a
given qemu run, you can reboot multiple times and still be ok.

However, if you stop qemu and re-run with the same disk image, it will
not be able to decrypt the rootfs and just reboots in a loop.

For now fix this by forcing a clean start each time.

Signed-off-by: Bill Mills <bill.mills@linaro.org>