build-qemu.sh: Build qemu 9.0.0-rc1
Using the system's qemu-system-aarch64 does not work for Zephyr under Xen.
Running a Zephyr DOMU will cause QEMU to hang 75% of the time.

This problem goes away by building qemu 9.0.0-rc1 (the latest at the time
of this work).

In order to have a build that will work on most people's machines, build
it on ubuntu 20.04.  This is the oldest Ubuntu not EOL from a public POV.

Right now I am using multipass to build it but it can also be run on
a clean vm or ec2 machine etc.  I will add docker support in the future.

Signed-off-by: Bill Mills <bill.mills@linaro.org>