Domu: config for Linux with Xen PV disk and network
I used this config (with gicv2) on TRS fine.
On debian I had to comment out the vif as debian does not
have xenbr0 right now

On both systems I created a small empty disk with:
dd if=/dev/zero of=disk0.img bs=1M count=16

I was able to use fdisk and mkfs.vat on the disk and then mount it.
The data persistet between runs and could be seem on the host with hexdump.

Signed-off-by: Bill Mills <bill.mills@linaro.org>