Testing: Record test logs from ec2 x86 and arm64 servers
This testing was done to prove that the qemu-9-wip baseline is solid and
not susceptible to the hang from earlier versions.

It is a bit out of date at the time of commit but keep a record for now.

Note: The hangs in the second x86 log are from switching to the system
provided qemu.  Note the QEMU_DIR=system.  This proves the same issue occurs on another machine.
It was not just a timing thing on my machine.

Signed-off-by: Bill Mills <bill.mills@linaro.org>