maybe-fetch: rename I to S.
When this was written, the saved-images were in "images" and thus I.
At that time the in use images were in generated and thus G.

Now in use images are in images so I is very confusing.
So change all $I's to $S's and all G's to U's (for in "use")
Avoid using $I for anything.

Signed-off-by: Bill Mills <bill.mills@linaro.org>