test-qemu: fix shellcheck errors
Fix a number of shellcheck errors to try to get it to work with dash or
ash.  However the use of a bash array for TESTS makes this hard and it
still won't parse in ash.

main won't run on the target but it still needs to parse.
So this goal failed but it is better to match std shell everywhere else.

Signed-off-by: Bill Mills <bill.mills@linaro.org>