zephyr-apps: Fix dom0less building
hello-mod needs an overlay to work correctly as dom0less.
Add this overlay to the command line for dom0less.
xen-irq-latency has also been changed to match this patern.
Update zephyr-apps to the one that has these changes
Update saved-images to the one that contains the results of running the
new build-zephyr-apps.sh

Signed-off-by: Bill Mills <bill.mills@linaro.org>